requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-extensions==4.1
djangorestframework==3.15.1
idna==3.7
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
pytest==8.2.0
pytest-django==4.8.0
requests==2.31.0
sqlparse==0.5.0
urllib3==2.2.1
